Crate massa_executed_ops

Copyright (c) 2022 MASSA LABS info@massa.net

Modules

config 🔒

Copyright (c) 2022 MASSA LABS info@massa.net

denunciations_changes 🔒

Copyright (c) 2023 MASSA LABS info@massa.net

executed_denunciations 🔒

Copyright (c) 2023 MASSA LABS info@massa.net This file defines a structure to list and prune previously executed denunciations. Used to detect denunciation reuse.

executed_ops 🔒

Copyright (c) 2022 MASSA LABS info@massa.net This file defines a structure to list and prune previously executed operations. Used to detect operation reuse.

ops_changes 🔒

Copyright (c) 2022 MASSA LABS info@massa.net

Macros

denunciation_index_key

Denunciation index key formatting macro

op_id_key

Op id key formatting macro

Structs

ExecutedDenunciations

A structure to list and prune previously executed denunciations

ExecutedDenunciationsChangesDeserializer

Deserializer for ExecutedOps

ExecutedDenunciationsChangesSerializer

ExecutedOps Serializer

ExecutedDenunciationsConfig

ExecutedOps

A structure to list and prune previously executed operations

ExecutedOpsChangesDeserializer

Deserializer for ExecutedOps

ExecutedOpsChangesSerializer

ExecutedOps Serializer

ExecutedOpsConfig

Type Aliases

ExecutedDenunciationsChanges

Speculative changes for ExecutedOps

ExecutedOpsChanges

Changes for ExecutedOps (was_successful, op_expiry_slot)